+++ tests/integrations/raftstore/mod.rs
@@ -18,7 +18,6 @@ mod test_region_heartbeat;
 mod test_region_info_accessor;
 mod test_replica_read;
 mod test_replication_mode;
-mod test_service;
 mod test_single;
 mod test_snap;
 mod test_split_region;
